moved Stepper state and control to RefugeeForm
