+++ packages/dla-piper/src/components/refugeeForm/refugeeForm.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { connect } from "frontity";
 import { Typography } from "@mui/material";
 import {
@@ -13,7 +13,6 @@ import { WhoAreYouStep, TravelStep, VisaStep, FamilyStep, AdditionalStep } from 
 const RefugeeForm = ({ state }) => {
   const data = state.source.get(state.router.link);
   const refugeeForm = state.source[data.type][data.id];
-
   const {
     rfTitle,
     rfDescription,
@@ -21,9 +20,32 @@ const RefugeeForm = ({ state }) => {
     rfInfoListItems,
   } = refugeeForm.acf;
 
-  useEffect(() => {
-    console.log("refugeeForm.acf", refugeeForm.acf);
-  }, []);
+  // stepper state etc...
+  const [currentStep, setCurrentStep] = useState(0);
+  const formConfig = [
+    { step: 1 },
+    { step: 2 },
+    { step: 3 },
+    { step: 4 },
+    { step: 5 },
+  ]
+  const handleNextStep = () => {
+    setCurrentStep(step => {
+      if (step === formConfig.length - 1) {
+        return step;
+      }
+      return step + 1;
+    })
+  }
+  const handlePreviousStep = () => {
+    setCurrentStep(step => {
+      if (step === 0) {
+        return step;
+      }
+      return step - 1;
+    })
+  }
+
 
   return (
     <>
@@ -64,14 +86,18 @@ const RefugeeForm = ({ state }) => {
           )}
         </MaxRestraintWrapper>
       </PageHeader>
+      
       <MaxRestraintWrapper>
-        <Stepper steps={[
-          <WhoAreYouStep />,
-          <TravelStep />,
-          <VisaStep />,
-          <FamilyStep />,
-          <AdditionalStep />,
-          ]} />
+        <Stepper 
+          currentStep={currentStep}
+          steps={[
+            <WhoAreYouStep onNext={handleNextStep} />,
+            <TravelStep onNext={handleNextStep} onPrevious={handlePreviousStep} />,
+            <VisaStep onNext={handleNextStep} onPrevious={handlePreviousStep} />,
+            <FamilyStep onNext={handleNextStep} onPrevious={handlePreviousStep} />,
+            <AdditionalStep onPrevious={handlePreviousStep} />,
+          ]} 
+        />
       </MaxRestraintWrapper>
     </>
   );
